Add files via upload
A gear is a wheel with teeth around its circumference. Gears are usually found in sets of two or more, used to transmit rotation from the axis of one gear to the axis of another. The teeth of a gear one one axis mesh with the teeth of a gear on another, thus creating a relationship between the rotation of the two axes. When one axis is spun, the other will too. Two gears of different sizes will make their two axes spin at different speeds, which you'll learn about, along with different types of gears and places they're used.